+++ src/main/java/com/elmakers/mine/bukkit/api/maps/URLMap.java
@@ -1,0 +1,8 @@
+package com.elmakers.mine.bukkit.api.maps;
+
+public interface URLMap {
+    public short getId();
+    public String getName();
+    public String getURL();
+    public boolean matches(String keyword);
+}